cp: refactor repetitions
