+++ src/com/facebook/buck/jvm/java/DefaultJavaLibrary.java
@@ -16,13 +16,8 @@
 
 package com.facebook.buck.jvm.java;
 
-import static com.facebook.buck.util.zip.ZipCompressionLevel.DEFAULT;
-import static com.facebook.buck.util.zip.ZipCompressionLevel.NONE;
-import static com.facebook.buck.util.zip.ZipOutputStreams.HandleDuplicates.APPEND_TO_ZIP;
-
 import com.facebook.buck.android.packageable.AndroidPackageable;
 import com.facebook.buck.android.packageable.AndroidPackageableCollector;
-import com.facebook.buck.io.BuildCellRelativePath;
 import com.facebook.buck.io.filesystem.ProjectFilesystem;
 import com.facebook.buck.jvm.core.HasClasspathDeps;
 import com.facebook.buck.jvm.core.HasClasspathEntries;
@@ -48,18 +43,9 @@ import com.facebook.buck.rules.SourcePathResolver;
 import com.facebook.buck.rules.SupportsPipelining;
 import com.facebook.buck.rules.keys.SupportsDependencyFileRuleKey;
 import com.facebook.buck.rules.keys.SupportsInputBasedRuleKey;
-import com.facebook.buck.step.ExecutionContext;
 import com.facebook.buck.step.Step;
-import com.facebook.buck.step.StepExecutionResult;
-import com.facebook.buck.step.StepExecutionResults;
-import com.facebook.buck.step.fs.CopyStep;
-import com.facebook.buck.step.fs.MakeCleanDirectoryStep;
-import com.facebook.buck.step.fs.RmStep;
 import com.facebook.buck.util.HumanReadableException;
 import com.facebook.buck.util.MoreSuppliers;
-import com.facebook.buck.util.zip.CustomZipEntry;
-import com.facebook.buck.util.zip.CustomZipOutputStream;
-import com.facebook.buck.util.zip.ZipOutputStreams;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
@@ -69,9 +55,7 @@ import com.google.common.collect.ImmutableSortedMap;
 import com.google.common.collect.ImmutableSortedSet;
 import com.google.common.collect.Sets;
 import com.google.common.hash.HashCode;
-import com.google.common.io.ByteStreams;
 import java.io.IOException;
-import java.io.InputStream;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Optional;
@@ -79,11 +63,6 @@ import java.util.Set;
 import java.util.SortedSet;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
-import java.util.jar.Attributes;
-import java.util.jar.JarFile;
-import java.util.jar.Manifest;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipInputStream;
 import javax.annotation.Nullable;
 
 /**
@@ -472,77 +451,9 @@ public class DefaultJavaLibrary extends AbstractBuildRule
       return steps.build();
     }
 
-    final Path outputJar = buildContext.getSourcePathResolver()
-        .getRelativePath(getSourcePathToOutput());
-
-    // Unzip the output, modify the manifest, rewrite the output
-    Path stampRoot = outputJar.getParent().resolve("temp-stamping");
-    Path stampedJar = stampRoot.resolve("stamped.jar");
-    steps.addAll(MakeCleanDirectoryStep.of(BuildCellRelativePath.of(stampRoot)));
-
-    steps.add(new Step() {
-      @Override
-      public StepExecutionResult execute(ExecutionContext context)
-          throws IOException, InterruptedException {
-
-        Manifest manifest = getProjectFilesystem().getJarManifest(outputJar);
-        if (manifest == null) {
-          manifest = new Manifest();
-        }
-
-        Attributes attrs = manifest.getEntries().getOrDefault("Build-Info", new Attributes());
-        attrs.putValue("Build-Revision", stamp.getSourceRevision());
-        attrs.putValue("Build-Time", stamp.getDatestamp());
-        attrs.putValue("Build-User", stamp.getUsername());
-        manifest.getEntries().put("Build-Info", attrs);
-
-        Path temp = stampRoot.resolve("stamped.jar");
-        try (
-            CustomZipOutputStream zos = ZipOutputStreams.newOutputStream(temp, APPEND_TO_ZIP);
-            InputStream is = getProjectFilesystem().newFileInputStream(outputJar);
-            ZipInputStream zis = new ZipInputStream(is)) {
-          // Write out our new manifest
-          CustomZipEntry entry = new CustomZipEntry("META-INF");
-          entry.setCompressionLevel(NONE.getValue());
-          zos.putNextEntry(entry);
-          zos.closeEntry();
-
-          entry = new CustomZipEntry(JarFile.MANIFEST_NAME);
-          entry.setCompressionLevel(DEFAULT.getValue());
-          zos.putNextEntry(entry);
-          manifest.write(zos);
-          zos.closeEntry();
-
-          // Now copy everything else, as is.
-          for (ZipEntry toCopy = zis.getNextEntry(); toCopy != null; toCopy = zis.getNextEntry()) {
-            if ("META-INF".equals(toCopy.getName()) ||
-                JarFile.MANIFEST_NAME.equals(toCopy.getName())) {
-              continue;
-            }
-
-            zos.putNextEntry(toCopy);
-            ByteStreams.copy(zis, zos);
-            zos.closeEntry();
-          }
-        }
-
-        return StepExecutionResults.SUCCESS;
-      }
-
-      @Override
-      public String getShortName() {
-        return "stamp";
-      }
-
-      @Override
-      public String getDescription(ExecutionContext context) {
-        return "Add build stamp";
-      }
-    });
-
-    steps.add(RmStep.of(BuildCellRelativePath.of(outputJar)));
-    steps.add(CopyStep.forFile(getProjectFilesystem(), stampedJar, outputJar));
-
+    steps.addAll(
+        ManifestBuildStamping.getBuildStampingSteps(
+            buildContext, stamp, getProjectFilesystem(), getSourcePathToOutput()));
     return steps.build();
   }
 }
